Fix eventListener bug
addEventListener, once added to an element is not removed until
element is replaced or explicitly removed. If more addEventListeners
are added on top of already existing event listeners on element, each
event will now register multiple times equal to the number of listeners
existing on that element, even if the event only triggered once.